refactor: removed utils from token-bridge-sdk
