+++ packages/arb-token-bridge-ui/src/hooks/useERC20L1Address.ts
@@ -1,6 +1,26 @@
 import { Provider } from '@ethersproject/providers'
 import useSWRImmutable from 'swr/immutable'
-import { getL1ERC20Address } from 'token-bridge-sdk'
+import { Erc20Bridger } from '@arbitrum/sdk'
+
+/**
+ * Retrieves the L1 address of an ERC-20 token using its L2 address.
+ * @param erc20L2Address
+ * @returns
+ */
+async function getL1ERC20Address({
+  erc20L2Address,
+  l2Provider
+}: {
+  erc20L2Address: string
+  l2Provider: Provider
+}): Promise<string | null> {
+  try {
+    const erc20Bridger = await Erc20Bridger.fromProvider(l2Provider)
+    return await erc20Bridger.getL1ERC20Address(erc20L2Address, l2Provider)
+  } catch (error) {
+    return null
+  }
+}
 
 /**
  * Returns L1 address
@@ -36,4 +56,4 @@ const useERC20L1Address = ({
   return { data, isLoading: isValidating }
 }
 
-export { useERC20L1Address }
+export { getL1ERC20Address, useERC20L1Address }
