--- packages/arb-token-bridge-ui/src/token-bridge-sdk/util/getL1ERC20Address.ts
@@ -1,24 +1,0 @@
-import { Provider } from '@ethersproject/providers'
-import { Erc20Bridger } from '@arbitrum/sdk'
-
-/**
- * Retrieves the L1 address of an ERC-20 token using its L2 address.
- * @param erc20L2Address
- * @returns
- */
-async function getL1ERC20Address({
-  erc20L2Address,
-  l2Provider
-}: {
-  erc20L2Address: string
-  l2Provider: Provider
-}): Promise<string | null> {
-  try {
-    const erc20Bridger = await Erc20Bridger.fromProvider(l2Provider)
-    return await erc20Bridger.getL1ERC20Address(erc20L2Address, l2Provider)
-  } catch (error) {
-    return null
-  }
-}
-
-export { getL1ERC20Address }
