+++ packages/arb-token-bridge-ui/src/util/TokenUtils.ts
@@ -5,16 +5,8 @@ import { BigNumber, constants } from 'ethers'
 import { Provider } from '@ethersproject/providers'
 import { Erc20Bridger, MultiCaller } from '@arbitrum/sdk'
 import { StandardArbERC20__factory } from '@arbitrum/sdk/dist/lib/abi/factories/StandardArbERC20__factory'
-import { EventArgs } from '@arbitrum/sdk/dist/lib/dataEntities/event'
-import { L2ToL1TransactionEvent } from '@arbitrum/sdk/dist/lib/message/L2ToL1Message'
-import { L2ToL1TransactionEvent as ClassicL2ToL1TransactionEvent } from '@arbitrum/sdk/dist/lib/abi/ArbSys'
-
-import { ERC20__factory, L1TokenData, L2TokenData } from '../index'
-
-export function assertNever(x: never, message = 'Unexpected object'): never {
-  console.error(message, x)
-  throw new Error('see console ' + message)
-}
+import { ERC20__factory } from '@arbitrum/sdk/dist/lib/abi/factories/ERC20__factory'
+import { L1TokenData, L2TokenData } from '../token-bridge-sdk/index'
 
 export const validateTokenList = (tokenList: TokenList) => {
   const ajv = new Ajv()
@@ -186,10 +178,4 @@ export async function getL2TokenData({
     balance: tokenData?.balance ?? constants.Zero,
     contract
   }
-}
-
-export function isClassicL2ToL1TransactionEvent(
-  event: L2ToL1TransactionEvent
-): event is EventArgs<ClassicL2ToL1TransactionEvent> {
-  return typeof (event as any).batchNumber !== 'undefined'
 }
